rdev

Enum Key

pub enum Key {

    Alt,
    AltGr,
    Backspace,
    CapsLock,
    ControlLeft,
    ControlRight,
    Delete,
    DownArrow,
    End,
    Escape,
    F1,
    F10,
    F11,
    F12,
    F2,
    F3,
    F4,
    F5,
    F6,
    F7,
    F8,
    F9,
    Home,
    LeftArrow,
    MetaLeft,
    MetaRight,
    PageDown,
    PageUp,
    Return,
    RightArrow,
    ShiftLeft,
    ShiftRight,
    Space,
    Tab,
    UpArrow,
    PrintScreen,
    ScrollLock,
    Pause,
    NumLock,
    BackQuote,
    Num1,
    Num2,
    Num3,
    Num4,
    Num5,
    Num6,
    Num7,
    Num8,
    Num9,
    Num0,
    Minus,
    Equal,
    KeyQ,
    KeyW,
    KeyE,
    KeyR,
    KeyT,
    KeyY,
    KeyU,
    KeyI,
    KeyO,
    KeyP,
    LeftBracket,
    RightBracket,
    KeyA,
    KeyS,
    KeyD,
    KeyF,
    KeyG,
    KeyH,
    KeyJ,
    KeyK,
    KeyL,
    SemiColon,
    Quote,
    BackSlash,
    IntlBackslash,
    KeyZ,
    KeyX,
    KeyC,
    KeyV,
    KeyB,
    KeyN,
    KeyM,
    Comma,
    Dot,
    Slash,
    Insert,
    KpReturn,
    KpMinus,
    KpPlus,
    KpMultiply,
    KpDivide,
    Kp0,
    Kp1,
    Kp2,
    Kp3,
    Kp4,
    Kp5,
    Kp6,
    Kp7,
    Kp8,
    Kp9,
    KpDelete,
    Function,
    Unknown(u32),

}

Key names based on physical location on the device Merge Option(MacOS) and Alt(Windows, Linux) into Alt Merge Windows (Windows), Meta(Linux), Command(MacOS) into Meta Characters based on Qwerty layout, donโt use this for characters as it WILL depend on the layout. Use Event.name instead. Key modifiers gives those keys a different value too. Careful, on Windows KpReturn does not exist, itโ s strictly equivalent to Return, also Keypad keys get modified if NumLock is Off and ARE pagedown and so on.

Variants

Alt

Alt key on Linux and Windows (option key on macOS)

AltGr

Backspace

CapsLock

ControlLeft

ControlRight

Delete

DownArrow

End

Escape

F1

F10

F11

F12

F2

F3

F4

F5

F6

F7

F8

F9

Home

LeftArrow

MetaLeft

also known as โwindowsโ, โsuperโ, and โcommandโ

MetaRight

also known as โwindowsโ, โsuperโ, and โcommandโ

PageDown

PageUp

Return

RightArrow

ShiftLeft

ShiftRight

Space

Tab

UpArrow

PrintScreen

ScrollLock

Pause

NumLock

BackQuote

Num1

Num2

Num3

Num4

Num5

Num6

Num7

Num8

Num9

Num0

Minus

Equal

KeyQ

KeyW

KeyE

KeyR

KeyT

KeyY

KeyU

KeyI

KeyO

KeyP

LeftBracket

RightBracket

KeyA

KeyS

KeyD

KeyF

KeyG

KeyH

KeyJ

KeyK

KeyL

SemiColon

Quote

BackSlash

IntlBackslash

KeyZ

KeyX

KeyC

KeyV

KeyB

KeyN

KeyM

Comma

Dot

Slash

Insert

KpReturn

KpMinus

KpPlus

KpMultiply

KpDivide

Kp0

Kp1

Kp2

Kp3

Kp4

Kp5

Kp6

Kp7

Kp8

Kp9

KpDelete

Function

Unknown(u32)